netfilter: delete the warning message
When a packet flood_forward in a bridge which has more than 2 ports.
For the first port, it will clone the skb. the cloned skb will share the
ct(struct nf_conn) which created in the nf_conntrack_in in the original
skb.when the cloned skb reaches nf_rtcache_forward,it will check if it
was not confirmed and there doesn't exist the rtcache extension in ct,
if true, it creates a new one and put it into the ct.

if this forward get dropped by other forward or postrouting filters,
then ct will not be get confirmed by nf_ipv4_confirm.

for the second port, it also clone the skb and refer to the shared ct,
it will reaches again nf_rtcache_forward, it will check if it was not
confirmed and there doesn't exist the rtcache extension in the ct,
because it was put in first port round, the second condition was false,
then warning was hit.

The warning can be hit in a normal case, so we should not warning here

Change-Id: Idccd63d569b73caa827941a350681c4a7f7672a2
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>