tracing: Record ftrace data to coherent memory
Presently, the ftrace data is saved to cacheable memory.
In some situations, the system reboots unexpectedly and
the ftrace data is not flushed from the cache and the
saved ftrace data is lost. In memory dumps collected in
these situations, the developer doesn't have enough data
to get an idea of which functions where being executed
just before the reboot. To overcome this, the ftrace
data is being recorded in coherent memory, hence the
information is available after the reboot even though
the cache is not flushed.

Please note that if this is turned on, the normal ftrace
data is not recorded. Having both seems to be affect
performance.

Change-Id: I954e90ace6b7c446dc06b39266f7a17891235bbf
Signed-off-by: Varadarajan Narayanan <varada@codeaurora.org>
(cherry picked from commit 76250694f4484c4fd975fb099f611fe6ced2e9de)
Signed-off-by: Prasanna Kumar Thoorvas Samyrao Muralidharan <quic_ptsm@quicinc.com>