net:qrtr Fix in qrtr_node_release.
Working case:
When the remoteproc is restarted, it initiates transfer by
sending hello packet to APPS and expects the same response from APPS.
Afterwards the APPS and Q6 would broadcast all of its services to
each other.This is the initial phase.

However in our case when the remoteproc is restarted, APPS receives
hello packet from remote. But its not sending hello packet to the
remote. Its dropped in qrtr layer itself. This message was not sent
to the remote processor. So when checked, observed that node->ep
was null which it should never be if qrtr_endpoint_regitser is a
success, which it was. So it meant something was holding reference
to the node and qrtr_node_release was not handled properly.
If node never gets released then it will never be removed from the
map and when APPS try to send hello packets the old reference
in the map would be looked which is null in our case.

This has the changes for fixing this scenario.

Signed-off-by: Nikhil Prakash V <nprakash@codeaurora.org>
(cherry picked from commit 9115334094c064e5ae2cbbbe4e713d5c2983214b)
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>

Change-Id: I0a6ce40443d44bfde45010afc17a24a296e5ed22