diag: usb: remove usb dependency from diag
If diag traffic over usb is required then USB_DIAG packages
should be added in the corresponding profile.

In diag drivers few parts of code are common for usb diag
& socket diag. Based on DIAG_OVER_USB config, segregated
the code so that we will be able to use socket diag
irrespective of whether USB is enabled or not.

Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>
(cherry picked from commit 8e438826ef394696850ac35597d7c5324eca7400)

Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>
Change-Id: I7ec54f50485441ee05faa97f3f5ffdcb5f740b05
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>