net: qrtr: Migrate nameservice to kernel from userspace
The QRTR nameservice has been maintained in userspace for some time. This
commit migrates it to Linux kernel. This change is required in order to
eliminate the need of starting a userspace daemon for making the WiFi
functional for ath11k based devices. Since the QRTR NS is not usually
packed in most of the distros, users need to clone, build and install it
to get the WiFi working. It will become a hassle when the user doesn't
have any other source of network connectivity.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 0c2204a4ad710d95d348ea006f14ba926e842ffd)

Change-Id: I0742244b27d77f7c5cd0dc5e35e310a6a2d85324
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>