firmware: qcom: scm: use device structure for DMA operations
In older kernel versions, if we don't use device structure for DMA opertions
for ARM32, it will take the dummy dma ops and DMA operations will succeed.
Now in the recent kernel version, device structure is mandatory for the
DMA operations, without it we will end up in unexpected behavior.

Update the DMA operations to take the device structure for successful
opertaion.

Change-Id: I8e0a23ce849f07963ad78fd5989a8cc4e196c38d
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>