OpenWrt: 030-modpost-add-a-helper-to-get-data-pointed-by-a-symbol.patch.
modpost: add a helper to get data pointed by a symbol

When CONFIG_MODULE_REL_CRCS is enabled, the value of __crc_* is not
an absolute value, but the address to the CRC data embedded in the
.rodata section.

Getting the data pointed by the symbol value is somewhat complex.
Split it out into a new helper, sym_get_data().

I will reuse it to refactor namespace_from_kstrtabns() in the next
commit.

Change-Id: I1f1d5216646de5bed0b537a0c25b09aa472ef34c
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>