OpenWrt: 040-ARM-8918-2-only-build-return_address-if-needed.patch.
ARM: 8918/2: only build return_address() if needed

The system currently warns if the config conditions for
building return_address in arch/arm/kernel/return_address.c
are not met, leaving just an EXPORT_SYMBOL_GPL(return_address)
of a function defined to be 'static linline'.
This is a result of aeea3592a13b ("ARM: 8158/1: LLVMLinux: use static
inline in ARM ftrace.h").

Since we're not going to build anything other than an exported
symbol for something that is already being defined to be an
inline-able return of NULL, just avoid building the code to
remove the following warning:

Fixes: aeea3592a13b ("ARM: 8158/1: LLVMLinux: use static inline in ARM
ftrace.h")

Change-Id: I21da85a2dea6c7b081998a915c88f5ff2f376a52
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>