OpenWrt: 717-v5.5-net-sfp-rework-upstream-interface.patch.
net: sfp: rework upstream interface

The current upstream interface is an all-or-nothing, which is
sub-optimal for future changes, as it doesn't allow the upstream driver
to prepare for the SFP module becoming available, as it is at boot.

Switch to a find-sfp-bus, add-upstream, del-upstream, put-sfp-bus
interface structure instead, which allows the upstream driver to
prepare for a module being available as soon as add-upstream is called.

Change-Id: I2f2ffbe6ffeff03f386b247ddb923128debd04cf
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>