OpenWrt: 726-v5.5-net-sfp-avoid-power-switch-on-address-change-modules.patch.
net: sfp: avoid power switch on address-change modules

If the module indicates that it requires an address change sequence to
switch between address 0x50 and 0x51, which we don't support, we can't
write to the register that controls the power mode to switch to high
power mode.  Warn the user that the module may not be functional in
this case, and don't try to change the power mode.

Change-Id: I91fc9756d3a9931889680a63142439bd63b57852
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>