OpenWrt: 727-v5.5-net-sfp-control-TX_DISABLE-and-phy-only-from-main-st.patch.
net: sfp: control TX_DISABLE and phy only from main
 state machine

We initialise TX_DISABLE when the sfp cage is probed, and then
maintain its state in the main state machine.  However, the module
state machine:
- negates it when detecting a newly inserted module when it's already
  guaranteed to be negated.
- negates it when the module is removed, but the main state machine
  will do this anyway.

Make TX_DISABLE entirely controlled by the main state machine.

The main state machine also probes the module for a PHY, and removes
the PHY when the the module is removed.  Hence, removing the PHY in
sfp_sm_module_remove() is also redundant, and is a left-over from
when we tried to probe for the PHY from the module state machine.

Change-Id: If1b1803293db8dd562e5fa21f6b56fb2ee8af2f7
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>