OpenWrt: 732-v5.5-net-sfp-track-upstream-s-attachment-state-in-state-m.patch.
net: sfp: track upstream's attachment state in state machine

Track the upstream's attachment state in the state machine rather than
maintaining a boolean, which ensures that we have a strict order of
ATTACH followed by an UP event - we can never believe that a newly
attached upstream will be anything but down.

Rearrange the order of state machines so we run the module state
machine after the upstream device's state machine, so the module state
machine can check the current state of the device and take action to
e.g. reset back to empty state when the upstream is detached.

This is to allow the module detection to run independently of the
network device becoming available.

Change-Id: I04ede7691c218038bce49b0cec2baed54d210293
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>