OpenWrt: 733-v5.5-net-sfp-split-power-mode-switching-from-probe.patch.
net: sfp: split power mode switching from probe

Switch the power mode switching from the probe, so that we don't
repeatedly re-probe the SFP device if there is a problem accessing
the registers at I2C address 0x51.

In splitting this out, we can also fix a bug where we leave the module
in high-power mode when the upstream device is detached but the module
is still inserted.

Change-Id: I1db4c49dfb6ec84c24161e18cafa6e9ba387c58d
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>