OpenWrt: 734-v5.5-net-sfp-move-module-insert-reporting-out-of-probe.patch.
net: sfp: move module insert reporting out of probe

Move the module insertion reporting out of the probe handling, but
after we have detected that the upstream has attached (since that is
whom we are reporting insertion to.)

Only report module removal if we had previously reported a module
insertion.

This gives cleaner semantics, and means we can probe the module before
we have an upstream attached.

Change-Id: Ifa580da7e888b1af9d679e0ffbee321a7675d876
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>