OpenWrt: 738-v5.5-net-phy-marvell10g-add-SFP-support.patch.
net: phy: marvell10g: add SFP+ support

Add support for SFP+ cages to the Marvell 10G PHY driver. This is
slightly complicated by the way phylib works in that we need to use
a multi-step process to attach the SFP bus, and we also need to track
the phylink state machine to know when the module's transmit disable
signal should change state.

With appropriate DT changes, this allows the SFP+ canges on the
Macchiatobin platform to be functional.

Change-Id: I6dba0fc8b0c77644d2824e6770aef01f9aeb497e
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>