OpenWrt: 309-MIPS-Add-CPU-option-reporting-to-proc-cpuinfo.patch
MIPS: Add CPU option reporting to /proc/cpuinfo

Many MIPS CPUs have optional CPU features which are not activates for
all CPU cores. Print the CPU options which are implemented in the core
in /proc/cpuinfo. This makes it possible to see what features are
supported and which are not supported. This should cover all standard
MIPS extensions, before it only printed information about the main MIPS
ASEs.

Change-Id: I8c46fb6b0889a70e5a59d3f385389fc47ddd0f64
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>