OpenWrt: 460-mtd-cfi_cmdset_0002-no-erase_suspend.patch
kernel: disable cfi cmdset 0002 erase suspend

on some platforms, erase suspend leads to data corruption and lockups when
write ops collide with erase ops. this has been observed on the buffalo
wzr-hp-g300nh. rather than play whack-a-mole with a hard to reproduce
issue on a variety of devices,simply disable erase suspend, as it will
usually not produce any useful gain on the small filesystems used
on embedded hardware.

Change-Id: I64890490512977baeb75f4d8fcbda1fb0aad81e3
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>