OpenWrt: 640-netfilter-nf_flow_table-add-hardware-offload-support.patch
nf_flow_table: add hardware offload support

This patch adds the infrastructure to offload flows to hardware, in case
the nic/switch comes with built-in flow tables capabilities.

If the hardware comes with no hardware flow tables or they have
limitations in terms of features, the existing infrastructure falls back
to the software flow table implementation.

The software flow table garbage collector skips entries that resides in
the hardware, so the hardware will be responsible for releasing this
flow table entry too via flow_offload_dead().

Hardware configuration, either to add or to delete entries, is done from
the hardware offload workqueue, to ensure this is done from user context
given that we may sleep when grabbing the mdio mutex.

Change-Id: I1d9daeba33652bb47e26c82fdcc9fab4ab2111b0
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>