OpenWrt: 750-net-phylink-delay-MAC-configuration-for-copper-SFP-m.patch
net: phylink: delay MAC configuration for copper SFP modules

Knowing whether we need to delay the MAC configuration because a module
may have a PHY is useful to phylink to allow NBASE-T modules to work on
systems supporting no more than 2.5G speeds.

This commit allows us to delay such configuration until after the PHY
has been probed by recording the parsed capabilities, and if the module
may have a PHY, doing no more until the module_start() notification is
called.  At that point, we either have a PHY, or we don't.

We move the PHY-based setup a little later, and use the PHYs support
capabilities rather than the EEPROM parsed capabilities to determine
whether we can support the PHY.

Change-Id: I5810018587c7e6a1881767a323f7137b826e02f5
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>