OpenWrt: 820-libata-Assign-OF-node-to-the-SCSI-device.patch
libata: Assign OF node to the SCSI device

When we spawn a SCSI device from an ATA device in libata-scsi
the SCSI device had no relation to the device tree.

The DT binding allows us to define port nodes under a
PATA (IDE) or SATA host controller, so we can have proper device
nodes for these devices.

If OF is enabled, walk the children of the host controller node
to see if there is a valid device tree node to assign. The reg
is used to match to ID 0 for the master device and ID 1 for the
slave device.

The corresponding device tree bindings have been accepted by
the device tree maintainers.

Change-Id: I6aaeeee3d563585be5fad2a1f8586dc9c788e22f
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>