OpenWrt: 834-ledtrig-libata.patch
libata: add ledtrig support

This adds a LED trigger for each ATA port indicating disk activity.

As this is needed only on specific platforms (NAS SoCs and such),
these platforms should define ARCH_WANTS_LIBATA_LEDS if there
are boards with LED(s) intended to indicate ATA disk activity and
need the OS to take care of that.
In that way, if not selected, LED trigger support not will be
included in libata-core and both, codepaths and structures remain
untouched.

Change-Id: Id88cd938a3e7427428a6ab9d36f707cb74ac46dc
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>