OpenWrt: 550-loop-Report-EOPNOTSUPP-properly.patch
loop: Report EOPNOTSUPP properly.

Properly plumb out EOPNOTSUPP from loop driver operations, which may
get returned when for instance a discard operation is attempted but not
supported by the underlying block device. Before this change, everything
was reported in the log as an I/O error, which is scary and not
helpful in debugging.

Change-Id: I652d47601f913b73424fc3ab9bd6fb0bb6d70f5d
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>