OpenWrt: 551-loop-Better-discard-support-for-block-devices.patch
loop: Better discard support for block devices.

If the backing device for a loop device is itself a block device,
then mirror the "write zeroes" capabilities of the underlying
block device into the loop device. Copy this capability into both
max_write_zeroes_sectors and max_discard_sectors of the loop device.

The reason for this is that REQ_OP_DISCARD on a loop device translates
into blkdev_issue_zeroout(), rather than blkdev_issue_discard(). This
presents a consistent interface for loop devices (that discarded data
is zeroed), regardless of the backing device type of the loop device.
There should be no behavior change for loop devices backed by regular
files.

This change fixes blktest block/003, and removes an extraneous
error print in block/013 when testing on a loop device backed
by a block device that does not support discard.

Change-Id: Ife7fbcec47066e8c57e2bbe413592c6fb5d5934d
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>