phy: qcom-qmp: add QPHY_POWER_DOWN_CONTROL offset in reg layouts
QPHY_POWER_DOWN_CONTROL macro was removed and change was
made to take the offset of the same from the reg layout
array. There were few layouts in which the offset was not
added. Thus wrong offset was added in these cases.

Added the QPHY_POWER_DOWN_CONTROL offset in the missed
reg layout arrays.

Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>
Change-Id: Idcf6ce398b30271ed926670770e5268c9554d966