soc: qcom: ramdump: avoid extra atomic vatiable
The purpose of atomic variable g_class_refcnt and spinlock g_rd_dev_lock
is to track the number of devices created by ramdump device and destroy
the class once all the devices are removed.

since the pupose of the variables are same we don't need to put them
in two different locks. We can use the spinlock to achieve the
functionality.

Suggested-by: Varadarajan Narayanan <varada@codeaurora.org>
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>
Change-Id: I7c96f14bdff7650d33d1ed2580da9b4944816e8f