drivers: pci: Change sys noc clock disabling order
This patch makes sure that sys noc clock is the last
clock to be disabled in the deinit function.

This will ensure proper working of PCIE because sysnoc
should be the first clock to be enabled and the last
clock to be disabled.

This change is referred from below commit:
b781c46701cfe9a25af1964c7c896279b20e35cc

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: I024d7998e3bf8d7353726e6af1d34c5996117fdb