arm64: dts: ipq8074: coresight: update the coresight name for CTI devices
Coresight device names are allocated dynamically during probe.
Coresight name for CTI are defined in DTS as "coresight-ctiX"
whereas during CTI probe, device names are allocated as "ctiX".
Due to this mismatch, probe failure observed for TMC ETR and ETF.

To avoid this, coresight-name for CTI are assigned with "ctiX",
in DTS as well.

Signed-off-by: Kathiravan T <kathirav@codeaurora.org>
Change-Id: Ie62b6ffe58a3e17e23280a60364c0e47e7dcd956