drivers: bus: mhi: Temp Fix for RDDM path
As part of mhi_pm_disable_transition(), the mhi_cntrl->ee
changes to MHI_EE_DISABLE_TRANSITION. When MHI reset is
performed BHI intvec is generated as part of MHI reset (clear).

It is calling status call back again because the below
if condition is true:
If (mhi_cntrl->ee == MHI_EE_RDDM && mhi_cntrl->ee != ee)

This patch adds a check based on mhi state to ignore the second
status_cb based on the mhi state.

Without this patch, we will get 2 status_cb's with RDDM as
reason to the mhi client and it will go into recovery path
2 times causing dual reboot issue in RDDM path.

This is a temporary fix which will be reverted later with
a more appropriate fix based on mhi dev state.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: Id6e56ad316712997b5ecf8b5d5d0d0e34b4b1c10