crypto: tcrypt: disable crypto for secure nand
Secure nand uses fixed key where the key from crypto manager is
ignored and secure key from tz is obtained instead.

This will lead to failures in crypto test algorithms.

Crypto tests are not valid when the fixed key is enabled in qce,
but there is no way that crypto manager will know if fixed key is enabled.

To resolve this issue, this patch disables tests for crypto algorithms
used by secure nand.

Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>
Change-Id: I68a4efc0fe2cc02583ea94536a3bcfcb2a15d9d0