soc: introduce config QTI_APSS_ACC
cpu_ops and cpu_pwr_ctl is required only for the stand alone boot
up of secondary cores. With the full boot chain, this driver is not
required since the PSCI is used. So put this driver under a config
and enable it only when required.

This change also fixes the compilation issues when the QTI_APSS_ACC
is enabled. Also removes the qcom_scm_set_cold_boot_addr call which
is a no-op.

Signed-off-by: Kathiravan T <kathirav@codeaurora.org>
Change-Id: I2d0c85e10c2592100e9de54ffeb08a30d70882a5