drivers: bus: mhi: Update Test driver to test multiple PCI MSI
This patch updates the mhi test driver to register a dummy handler
for the irq vectors 3 to 127 to test the same. Vectors 0-2 will be used
by mhi bus driver.

In this patch, we assign 128 vectors from 4 MSI controllers and also
add support to deinit the vectors assigned so pci removal path is fine.

Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: I5fdee451f006657c74957507deb11fbb8c479d31