drivers: pci: qcom: Add naming for multiple PCI IRQ
Added naming for PCI IRQ, based on the PCIE domain number
and MSI controller number for the purpose of setting affinity
while used by the CNSS driver scripts.

In other words, this patch basically associates a name
with each of the GIC irq lines so the same can be used
while setting affinity.

Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: I0f44a9f8cb7fa895d945ec34545acd1df7eff51e