diag: always listen from the server once it starts running
currently diag data and dci servers are not running until the
pheripheral services are up. This causes loss in packet
if a pheripheral is not available. Start listening from a diag server
as soon as it is created to avoid losing any packets.

eggplant commit: cb5bfbe2cf7a54045675052d22057f02cb36eec3

Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>
Change-Id: I3cbc8b9114b6b0920cfb0454d9b7710e3c2384af