ipq807x: diag: do not close TYPE_CMD channel
There is an issue while doing wifi down/up that commands not being sent
from QXDM to Q6 subsystem.

It is actually reaching till diag_socket_writebut because of cleaning
CMD type during wifi down enough information was not available to pass
further. Hence the it gets failed.

With this patch we are not releaseing  the CMD channels.

Signed-off-by: Tushar Nimkar <tnimkar@codeaurora.org>
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>
Change-Id: I5cdee006879aa6d45f0ee7a4959e68f9ad2e523e