mtd: nand: qcom: Proper DMA sync for register read buffer
DMA mapping support is added for reg_read_buf. During the start
of any operation, this buffer will be synced for device operation.
After operation completion, it will be synced again for CPU.

In check_flash_errors, we are directly checking the reg_read_buf
without doing a DMA sync for CPU operation. This leads to incorrect
or older copy of DMA buffer in reg_read_buf. So, before checking the
flash status reg_read_buf needs to be synced for CPU operation. This
ensures the up-to-date and correct copy of DMA buffer in reg_read_buf.

This patch adds the DMA sync on reg_read_buf for CPU operation in
check_flash_errors function.

Signed-off-by: Praveenkumar I <ipkumar@codeaurora.org>
Change-Id: I9e1715eb73636202159eb039edc0c36539576373