net: cnss2: Add MAX_TGT_MEM_MODES macro
Add MAX_TGT_MEM_MODES macro with value 5 and use this instead of hard
coding the mem modes to 3 in multiple places.

Also add dummy entries for BDF and Caldb-address for modes 3 and 4
for IPQ6018 and IPQ5018 and define the same values as mode 0 for
modes 3 and 4 for IPQ8074.

The modes 3 and 4 for IPQ8074 would be defined in the FW but would
not be checked in host and can be enabled if required by modifying the
DTS file.

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: Ic33c3e51911d21a7494cf5099db15b5737a0a954