remoteproc: wcss: address possible array index out of bound
while collecting subsystem crashdump, the number of segments to dump
is determined from device tree.

if the number of segments mentioned in device tree exceeds MAX segments
size there will be a chance of exceeding array index.

this is resolved by fixing the number of segments to the minimum value
of segments mentioned in device tree and MAX segments allowed.

Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>
Change-Id: I296a7cc3884f898d82b141aa99395810f925e981